backend/services/notification-service/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.4
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
python-decouple==3.8
twilio==8.12.0
Pillow==10.2.0
